refactor(mobile): consolidate cubit tests into module_core
Move session_list and project_list cubit tests from app/test/features/
to module_core/test/cubits/ where the cubits are defined. Tests for a
component should live in the same Dart module as the component.

- Move session_list_cubit_test.dart to module_core
- Merge app/ and module_core/ project_list_cubit_test.dart into one
- Create module_core/test/helpers/test_helpers.dart with shared
  pure-Dart mocks and data factories
- Replace flutter_test imports with test (pure Dart package)
- Fix all analysis hints (import ordering, const, unused imports)
